feat(new-blog): queryString util 함수로 변경
