+++ src/pages/Posts.tsx
@@ -4,16 +4,17 @@ import styled from '@emotion/styled';
 import { useLocation } from 'react-router-dom';
 import { useAtom } from 'jotai';
 import { filteredPostsAtom, pagePostAtom, postsAtom } from '@store/posts';
+import { getQueryString } from '@utils/Utility';
 
 // components
 import Text from '@components/atoms/Text';
 import Search from '@components/molecules/posts/Search';
 import PostContents from '@components/molecules/posts/PostContents';
 
+// interfaces
 interface IQueryString {
   search?: string;
 }
-
 export default function Posts() {
   const location = useLocation();
   const [posts] = useAtom(postsAtom);
@@ -21,19 +22,6 @@ export default function Posts() {
   const [filteredPosts] = useAtom(filteredPostsAtom);
 
   const [searchInput, setSearchInput] = useState<string>('');
-
-  const getQueryString = (search: string) => {
-    const queryString = search.substring(1);
-    let queryInstance: IQueryString = {};
-
-    queryString.split('&').forEach((query) => {
-      const key = query.split('=')[0];
-      const value = query.split('=')[1];
-      queryInstance[key] = decodeURIComponent(value);
-    });
-
-    return queryInstance;
-  };
 
   const title = useMemo(() => (searchInput.length === 0 ? 'Post' : 'Search Result'), [searchInput]);
   const contentsData = useMemo(
@@ -43,7 +31,7 @@ export default function Posts() {
 
   // 쿼리스트링 처리
   useEffect(() => {
-    const queryInstance = getQueryString(location.search);
+    const queryInstance: IQueryString = getQueryString(location.search);
     if (queryInstance?.search) {
       setSearchInput(queryInstance?.search);
     }
